+++ pipeline/aggregate/blocks/emojis/EmojiStats.ts
