+++ pipeline/aggregate/blocks/interaction/ConversationStats.ts
