+++ pipeline/aggregate/blocks/interaction/InteractionStats.ts
