+++ pipeline/aggregate/blocks/sentiment/SentimentPerCycle.ts
