+++ pipeline/aggregate/blocks/sentiment/SentimentStats.ts
