+++ pipeline/aggregate/blocks/timeline/Growth.ts
