+++ report/components/cards/sentiment/SentimentOverTime.tsx
@@ -10,9 +10,9 @@ import {
     XYChart,
     XYCursor,
 } from "@amcharts/amcharts5/xy";
-import { SentimentInDate, SentimentPerCycle } from "@pipeline/aggregate/blocks/SentimentPerCycle";
+import { SentimentInDate, SentimentPerCycle } from "@pipeline/aggregate/blocks/sentiment/SentimentPerCycle";
 
-import { Themes, enableDebouncedResize, syncAxisWithTimeFilter } from "./AmCharts5";
+import { Themes, enableDebouncedResize, syncAxisWithTimeFilter } from "../../viz/amcharts/AmCharts5";
 
 const SentimentOverTime = ({ data, options }: { data?: SentimentPerCycle; options: number[] }) => {
     const chartDiv = useRef<HTMLDivElement>(null);
