+++ report/components/cards/timeline/ActiveAuthorsOverTime.tsx
@@ -1,21 +1,20 @@
 import { useLayoutEffect, useRef } from "react";
 
-import { Color, Label, Root, p50 } from "@amcharts/amcharts5";
-import { TimeUnit } from "@amcharts/amcharts5/.internal/core/util/Time";
+import { Bullet, Circle, Color, Label, Root, Tooltip, p50 } from "@amcharts/amcharts5";
 import {
     AxisRendererX,
     AxisRendererY,
     DateAxis,
-    StepLineSeries,
+    SmoothedXLineSeries,
     ValueAxis,
     XYChart,
     XYSeries,
 } from "@amcharts/amcharts5/xy";
-import { TimelineStats } from "@pipeline/aggregate/blocks/Growth";
+import { ActiveAuthors } from "@pipeline/aggregate/blocks/timeline/ActiveAuthors";
 
-import { Themes, enableDebouncedResize, syncAxisWithTimeFilter } from "./AmCharts5";
+import { Themes, enableDebouncedResize, syncAxisWithTimeFilter } from "../../viz/amcharts/AmCharts5";
 
-const GrowthOverTime = ({ data, options }: { data?: TimelineStats; options: number[] }) => {
+const ActiveAuthorsOverTime = ({ data, options }: { data?: ActiveAuthors; options: number[] }) => {
     const chartDiv = useRef<HTMLDivElement>(null);
     const xAxisRef = useRef<DateAxis<any> | null>(null);
     const seriesRef = useRef<XYSeries | null>(null);
@@ -25,7 +24,11 @@ const GrowthOverTime = ({ data, options }: { data?: TimelineStats; options: numb
         root.setThemes(Themes(root, false)); // Do not animate!
         const cleanupDebounce = enableDebouncedResize(root);
 
-        const chart = root.container.children.push(XYChart.new(root, {}));
+        const chart = root.container.children.push(
+            XYChart.new(root, {
+                layout: root.verticalLayout,
+            })
+        );
         chart.zoomOutButton.set("forceHidden", true);
 
         const xAxis = chart.xAxes.push(
@@ -44,40 +47,53 @@ const GrowthOverTime = ({ data, options }: { data?: TimelineStats; options: numb
         yAxis.children.unshift(
             Label.new(root, {
                 rotation: -90,
-                text: "Total authors",
+                text: "Active authors in period",
                 y: p50,
                 centerX: p50,
             })
         );
+        xAxisRef.current = xAxis;
 
-        const stepSeries = chart.series.push(
-            StepLineSeries.new(root, {
+        const series = chart.series.push(
+            SmoothedXLineSeries.new(root, {
                 valueXField: "ts",
                 valueYField: "value",
                 xAxis: xAxis,
                 yAxis: yAxis,
-                stroke: Color.fromHex(0x479adb),
-                fill: Color.fromHex(0x479adb),
+                stroke: Color.fromHex(0x57b1ff),
+                fill: Color.fromHex(0x57b1ff),
+                tooltip: Tooltip.new(root, {
+                    labelText: "{valueY}",
+                }),
             })
         );
-        stepSeries.strokes.template.setAll({
-            visible: true,
-            strokeWidth: 2,
-            strokeOpacity: 0.8,
-        });
-        stepSeries.fills.template.setAll({
+        series.fills.template.setAll({
             visible: true,
             fillOpacity: 0.2,
+            templateField: "lineSettings",
         });
+        series.strokes.template.setAll({
+            templateField: "lineSettings",
+        });
+        series.bullets.push(() =>
+            Bullet.new(root, {
+                locationY: 0,
+                sprite: Circle.new(root, {
+                    radius: 4,
+                    stroke: Color.fromHex(0x57b1ff),
+                    strokeWidth: 2,
+                    fill: Color.fromHex(0x1861a1),
+                }),
+            })
+        );
 
-        xAxisRef.current = xAxis;
-        seriesRef.current = stepSeries;
+        seriesRef.current = series;
 
         const cleanup = syncAxisWithTimeFilter([seriesRef.current], xAxis, yAxis);
 
         return () => {
-            cleanup();
             cleanupDebounce();
+            cleanup();
             root.dispose();
             xAxisRef.current = null;
             seriesRef.current = null;
@@ -85,13 +101,15 @@ const GrowthOverTime = ({ data, options }: { data?: TimelineStats; options: numb
     }, []);
 
     useLayoutEffect(() => {
-        xAxisRef.current?.set("baseInterval", {
-            timeUnit: "day" as TimeUnit,
-            count: 1,
-        });
         if (data) {
-            // TODO: update efficient
-            seriesRef.current?.data.setAll(data.growth);
+            if (data.perMonth.length > 1) {
+                // @ts-ignore
+                data.perMonth[data.perMonth.length - 2].lineSettings = {
+                    strokeDasharray: [3, 3],
+                    fillOpacity: 0.1,
+                };
+            }
+            seriesRef.current?.data.setAll(data.perMonth);
         }
     }, [seriesRef.current, data]);
 
@@ -107,4 +125,4 @@ const GrowthOverTime = ({ data, options }: { data?: TimelineStats; options: numb
     );
 };
 
-export default GrowthOverTime;
+export default ActiveAuthorsOverTime;
